+++ modules/client/shared/src/test/scala/com/azavea/stac4s/api/client/SttpStacClientFSpec.scala
@@ -16,13 +16,13 @@ import sttp.client3.testing.SttpBackendStub
 import sttp.model.Method
 import sttp.monad.EitherMonad
 
-trait SttpStacClientFSpec extends AnyFunSpec with Matchers with BeforeAndAfterAll {
+trait SttpStacClientFSpec[S] extends AnyFunSpec with Matchers with BeforeAndAfterAll {
 
   def arbCollectionShort: Arbitrary[StacCollection]
   def arbItemCollectionShort: Arbitrary[ItemCollection]
   def arbItemShort: Arbitrary[StacItem]
 
-  def client: SttpStacClientF[Either[Throwable, *]]
+  def client: SttpStacClientF[Either[Throwable, *], S]
 
   lazy val backend =
     SttpBackendStub(EitherMonad)