+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Update StacClient and SearchFilters JSON codecs [#305](https://github.com/azavea/stac4s/pull/305)
 
 ## [0.2.3] - 2021-05-04
 ### Fixed